Thought of how to implement next steps in the project. Working on profile settings

+++ backend/src/models/userModel.ts
@@ -7,11 +7,7 @@ import { removeUserColumns } from '../utils/functions'
 import { User, IDefault, IGetUserAndProjects, IUser, IUserAndProjects } from './types'
 
 export const getUserAndProjects = async (userId: string): Promise<IGetUserAndProjects | AppError> => {
-	const {
-		data: userWithProjects,
-		error,
-		status,
-	} = await supabase.from('users').select(`*, projects(*)`).eq('id', userId).single()
+	const { data: userWithProjects, error, status } = await supabase.from('users').select(`*, projects(*)`).eq('id', userId).single()
 
 	if (!userWithProjects) return new AppError(404)
 	if (error) return new AppError(status)
@@ -35,19 +31,20 @@ export const getUser = async (userId: string): Promise<IUser | AppError> => {
 }
 
 export const updateUser = async (reqBody: UpdateUserType, userId: string): Promise<IUser | AppError> => {
-	const response = await supabase.from('users').update(reqBody).eq('id', userId).select('*').single()
-	const { data: user, error, status } = response
+	// const response = await supabase.from('users').update(reqBody).eq('id', userId).select('*').single()
+	// const { data: user, error, status } = response
 
-	if (error) return new AppError(status)
-	if (!user) return new AppError(400)
+	// if (error) return new AppError(status)
+	// if (!user) return new AppError(400)
 
-	const newUser = removeUserColumns<User>(user)
+	// const newUser = removeUserColumns<User>(user)
 
-	return {
-		user: newUser,
-		statusCode: 200,
-		statusText: ['update', 'Your profile information has been updated successfully'],
-	}
+	// return {
+	// 	user: newUser,
+	// 	statusCode: 200,
+	// 	statusText: ['update', 'Your profile information has been updated successfully'],
+	// }
+	return new AppError(429)
 }
 
 export const deleteUser = async (password: string, userId: string): Promise<IDefault | AppError> => {
@@ -56,12 +53,7 @@ export const deleteUser = async (password: string, userId: string): Promise<IDef
 	const arePasswordsEqual = await bcrypt.compare(password, user.password)
 	if (!arePasswordsEqual) return new AppError(401, `Hmm, your passwords don't match. Try again.`)
 
-	const { error, status, data } = await supabase
-		.from('users')
-		.delete()
-		.eq('id', userId)
-		.select('fullName')
-		.single()
+	const { error, status, data } = await supabase.from('users').delete().eq('id', userId).select('fullName').single()
 
 	if (error) return new AppError(status)
